Python script to train and test the Text Generation DNN.
The Python script contains the Training and Testing subroutines to train and develop a LSTM based text generation DNN using PyTorch. The script can be toggled between Train and Test modes using the 'train' flag in the 'main'. 

The DNN is composed of a LSTM followed by a dense layer. The input file and parameters for training the can be altered in the Hyper-parameter settings.

Reference:
1. Text generation with PyTorch -   https://machinetalk.org/2019/02/08/text-generation-with-pytorch/