Python Source for training a W-GAN.
This Python source script is a GAN based Fake Image generator written using PyTorch. Models are defined in the Model_GAN.py source script.

The training data can be a set of 'jpg' or 'png' images.

The images will be resized before processing and the output images are of size 128 x 128 x CHANNEL, where CHANNEL is 3 for 'jpg' and 4 for 'png'. 

Training on a GPU is recommended.